testing: don't use expensive scrypt-65536-8-1 in integration tests
