+++ repo/crypto_key_derivation.go
@@ -4,25 +4,8 @@ import (
 	"crypto/sha256"
 	"io"
 
-	"github.com/pkg/errors"
 	"golang.org/x/crypto/hkdf"
-	"golang.org/x/crypto/scrypt"
 )
-
-// defaultKeyDerivationAlgorithm is the key derivation algorithm for new configurations.
-const defaultKeyDerivationAlgorithm = "scrypt-65536-8-1"
-
-func (f *formatBlob) deriveMasterKeyFromPassword(password string) ([]byte, error) {
-	const masterKeySize = 32
-
-	switch f.KeyDerivationAlgorithm {
-	case "scrypt-65536-8-1":
-		return scrypt.Key([]byte(password), f.UniqueID, 65536, 8, 1, masterKeySize)
-
-	default:
-		return nil, errors.Errorf("unsupported key algorithm: %v", f.KeyDerivationAlgorithm)
-	}
-}
 
 // deriveKeyFromMasterKey computes a key for a specific purpose and length using HKDF based on the master key.
 func deriveKeyFromMasterKey(masterKey, uniqueID, purpose []byte, length int) []byte {
